add 2 tests to SessionManagerTests

+++ server/src/main/java/dpas/server/session/SessionManager.java
@@ -44,7 +44,7 @@ public class SessionManager {
     /**
      * Validates an hmac for a valid session
      */
-    public void validateSessionRequest(String keyId, byte[] hmac, byte[] content, int sequenceNumber) throws NoSuchPaddingException, NoSuchAlgorithmException, InvalidKeyException, BadPaddingException, IllegalBlockSizeException {
+    public void validateSessionRequest(String keyId, byte[] hmac, byte[] content, long sequenceNumber) throws NoSuchPaddingException, NoSuchAlgorithmException, InvalidKeyException, BadPaddingException, IllegalBlockSizeException {
         if (!_sessionKeys.containsKey(keyId))
             throw new IllegalArgumentException("Invalid SessionId");
